% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp.R
\name{interp}
\alias{interp}
\title{Interpolates between Sets of Potential Outcomes}
\usage{
interp(Tr, y, tc1, tc2)
}
\arguments{
\item{Tr}{The treatment assignment vector.}

\item{y}{A vector of experimental outcomes.}

\item{tc1}{A matrix with 2 columns. The first column contains a set of imputed treatment outcomes. The second column contains a set of imputed control outcomes.}

\item{tc2}{A second set of imputed potential outcomes.}
}
\description{
This function is used within the \code{reloop} and \code{reloop_slow} imputation methods for the \code{loop} function. 
Using the input data as a training set, this function returns two weights (one for treated units and one for control units) 
that minimizes the mean squared error between the observed outcomes and a weighted average of two sets of potential outcomes.
The weight is constrained to be between 0 and 1.
}
