% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_ols.R
\name{loo_ols}
\alias{loo_ols}
\title{Leave-One-Out Linear Regression}
\usage{
loo_ols(Y, X)
}
\arguments{
\item{Y}{A vector of outcomes.}

\item{X}{A matrix of predictors.}
}
\description{
This function is used within the \code{loop_ols} imputation method for the \code{loop} function.
Given a set of outcomes \code{Y} and predictors \code{X}, this function returns the coefficients 
resulting from leaving each observation out and fitting ordinary least squares regression to the remaining observations.
}
