% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_fn.R
\name{loop}
\alias{loop}
\title{LOOP Estimator}
\usage{
loop(Y, Tr, Z, p = 0.5, dropobs = NULL)
}
\arguments{
\item{Y}{A vector of experimental outcomes.}

\item{Tr}{The treatment assignment vector.}

\item{Z}{A matrix or vector of covariates.}

\item{p}{The probability of being assigned to treatment. Defaults to 0.5.}

\item{dropobs}{When set to TRUE, lowers the bootstramp sample size in the random forest by 1 when making out-of-bag predictions. By default, this is set to TRUE if the sample size is less than or equal to 30 and FALSE otherwise.}
}
\description{
Covariate-adjusted estimator of the average treatment effect in randomized experiments. Given a set of experimental data (including observed outcome, treatment assignment, and covariates), uses random forests to impute the potential outcomes for each observation, which are then used to estimate the average treatment effect.
}
\examples{
## Create Simulated Data
N = 30 # 30 observations
k = 5 # 5 covariates

Z = matrix(runif(N*k)*10,N,k)
c = Z \%*\% runif(k) + runif(N)*5
t = ifelse(rowMeans(Z) < 3, c + runif(N)*5, c + runif(N)*10)

Tr = rbinom(N,1,0.5)
Y = ifelse(Tr == 0, c, t)

## Run LOOP and compare with a Difference in Means
loop.results = loop(Y, Tr, Z)

meandiff = mean(Y[Tr == 1]) - mean(Y[Tr == 0])
varhat = var(Y[Tr==1])/length(Y[Tr==1]) + var(Y[Tr==0])/length(Y[Tr==0])

print(c("loop",loop.results))
print(c("Difference in Means",meandiff,varhat))
}
