% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_ols.R
\name{loop_ols}
\alias{loop_ols}
\title{OLS Imputation of Potential Outcomes}
\usage{
loop_ols(Y, Tr, Z)
}
\arguments{
\item{Y}{A vector of observed outcomes.}

\item{Tr}{The treatment assignment vector.}

\item{Z}{A matrix of pre-treatment covariates.}
}
\description{
This function is used to impute potential outcomes for the \code{loop} function. 
Leaves out each observation and imputes its potential outcomes using ordinary least squares
regression on the remaining observations.
}
