% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_interp.R
\name{p_interp}
\alias{p_interp}
\title{Interpolates between Sets of Potential Differences in Paired Experiments}
\usage{
p_interp(Tr, a, b, ab1, ab2)
}
\arguments{
\item{Tr}{The treatment assignment vector for the pairs, where the value is 1 if the first unit in the pair is assigned to treatment and 0 otherwise.}

\item{a}{A vector of the treatment minus control differences for the pairs where the first unit is assigned to treatment.}

\item{b}{A vector of the treatment minus control differences for the pairs where the second unit is assigned to treatment.}

\item{ab1}{A matrix with 2 columns. The first column contains a set of imputed differences had the first unit in each pair been treated.  The first column contains a set of imputed differences had the second unit in each pair been treated.}

\item{ab2}{A second set of imputed potential differences.}
}
\description{
This function is used within the \code{p_rf_interp} and \code{p_ols_interp} imputation methods for the \code{p_loop} function. 
Given two sets of imputed potential differences, this function returns a single set of predicted differences by picking weights
that minimizes the mean squared error between the observed differences and the weighted average imputed differences. This minimization
is done using leave-one-out cross validation.
}
