% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_loop_var.R
\name{p_loop_var}
\alias{p_loop_var}
\title{Variance Estimate for the P-LOOP Estimator}
\usage{
p_loop_var(assigned, a, b)
}
\arguments{
\item{assigned}{A matrix of pair experimental data that his been processed by the \code{pair} function}

\item{a}{A vector of the treatment minus control differences for the pairs where the first unit is assigned to treatment.}

\item{b}{A vector of the treatment minus control differences for the pairs where the second unit is assigned to treatment.
#' @export}
}
\description{
This function implements the variance estimate for the P-LOOP estimator and is used within \code{p_loop}.
}
