% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_ols_ab.R
\name{p_ols_ab}
\alias{p_ols_ab}
\title{OLS Imputation of Potential Outcomes in Paired Experiments}
\usage{
p_ols_ab(assigned)
}
\arguments{
\item{assigned}{A matrix of pair experimental data that his been processed by the \code{pair} function.}
}
\description{
This function is used to impute potential outcomes for the \code{p_loop} function. 
Leaves out each pair and imputes its potential outcomes using ordinary least squares
regression on the remaining pairs. Pairs are treated as units when making predictions.
}
