% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_rf_ab.R
\name{p_rf_ab}
\alias{p_rf_ab}
\title{Random Forest Imputation of Potential Outcomes in Paired Experiments}
\usage{
p_rf_ab(assigned, reparam = TRUE)
}
\arguments{
\item{assigned}{A matrix of pair experimental data that his been processed by the \code{pair} function.}

\item{reparam}{If set to \code{TRUE}, covariates will be parameterized as the pairwise differences and means for each covariate.}
}
\description{
This function is used to impute potential outcomes for the \code{p_loop} function. 
Leaves out each pair and imputes its potential outcomes using random forests
on the remaining pairs. Pairs are treated as units when making predictions.
}
