% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_rf_interp.R
\name{p_rf_interp}
\alias{p_rf_interp}
\title{Random Forest Imputation of Potential Outcomes in Paired Experiments}
\usage{
p_rf_interp(assigned, reparam = TRUE)
}
\arguments{
\item{assigned}{A matrix of pair experimental data that his been processed by the \code{pair} function.}

\item{reparam}{If set to \code{TRUE}, covariates will be parameterized as the pairwise differences and means for each covariate when treating the pairs as units.}
}
\description{
This function is used to impute potential outcomes for the \code{p_loop} function. 
Leaves out each pair and imputes its potential outcomes using random forests
on the remaining pairs. This function interpolates between the \code{p_rf_ab} and \code{p_rf_s} methods.
}
