% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_rf_s.R
\name{p_rf_s}
\alias{p_rf_s}
\title{Random Forest Imputation of Potential Outcomes in Paired Experiments}
\usage{
p_rf_s(assigned)
}
\arguments{
\item{assigned}{A matrix of pair experimental data that his been processed by the \code{pair} function}
}
\description{
This function is used to impute potential outcomes for the \code{p_loop} function. 
Leaves out each pair and imputes its potential outcomes using random forests
on the remaining pairs. Individuals are treated as units when making predictions.
}
