% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair.R
\name{pair}
\alias{pair}
\title{Format Data into Pairs}
\usage{
pair(Y, Tr, Z, P)
}
\arguments{
\item{Y}{A vector of experimental outcomes.}

\item{Tr}{The treatment assignment vector.}

\item{Z}{A matrix or vector of covariates.}

\item{P}{A vector encoding the pair assignments. Observations with the same value of \code{P} will be assumed to be in the same pair.}
}
\description{
This function takes data at the individual level and returns a data frame with one pair per row. Each row includes the pair label,
the observed outcomes, the treatment assignment for the first individual in the pair, and the covariates for each individual.
}
