% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reloop.R
\name{reloop}
\alias{reloop}
\title{Imputation of Potential Outcomes using External Data}
\usage{
reloop(Y, Tr, Z, yhat)
}
\arguments{
\item{Y}{A vector of observed outcomes.}

\item{Tr}{The treatment assignment vector.}

\item{Z}{A matrix of pre-treatment covariates.}

\item{yhat}{A vector of external predictions.}
}
\description{
This function implements the ReLOOP method for imputing potential outcomes within the \code{loop} function. 
This method uses both covariates \code{Z} and external predictions \code{yhat} to impute potential outcomes, 
which can then be used within the LOOP estimator to obtain an estimate for the average treatment effect.
As implemented, \code{reloop} is very slightly biased. \code{reloop_slow} removes that bias at the cost of
computation time.
}
