% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorient.R
\name{reorient}
\alias{reorient}
\title{Format Paired Data}
\usage{
reorient(assigned, reparam = FALSE)
}
\arguments{
\item{assigned}{A matrix of pair experimental data that his been processed by the \code{pair} function.}

\item{reparam}{If set to \code{TRUE}, covariates will be parameterized as the pairwise differences and means for each covariate.}
}
\description{
This function takes a data frame that has been processed by the \code{pair} function and formats it such that the treated unit comes first
in each row.
}
