% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparam.R
\name{reparam}
\alias{reparam}
\title{Reparameterize Paired Data}
\usage{
reparam(assigned)
}
\arguments{
\item{assigned}{A matrix of pair experimental data that his been processed by the \code{pair} function.}
}
\description{
This function takes a data frame that has been processed by the \code{pair} function and formats it such that 
covariates are parameterized as the within pair differences and means for each covariate.
}
